Updates function URL retrieval for Gen 2
Updates the function URL retrieval method in the deployment workflow to be compatible with Cloud Functions Gen 2.

The previous method attempted to retrieve the URL from `httpsTrigger.url` or `uri`, which are not applicable to Gen 2 functions. This change updates the retrieval to use `serviceConfig.uri`, which is the correct location for Gen 2 function URLs.